Freeze datafile while making changes
