+++ tests/files/mine-out.yml
@@ -1,37 +1,37 @@
 config:
   computers:
-  - name: macbook
-    hostname: Jaces-MacBook
-    address: AA:BB:CC:DD:EE:FF
-    serial: abc123
-  - name: macbook-pro
-    hostname: Jaces-MacBook-2
-    address: 11:22:33:44:55:66
-    serial: def456
+    - name: macbook
+      hostname: Jaces-MacBook
+      address: AA:BB:CC:DD:EE:FF
+      serial: abc123
+    - name: macbook-pro
+      hostname: Jaces-MacBook-2
+      address: 11:22:33:44:55:66
+      serial: def456
   applications:
-  - name: iphoto
-    properties:
-      auto_queue: false
-      single_instance: false
-    versions:
-      mac: iPhoto
-      windows:
-      linux:
-  - name: itunes
-    properties:
-      auto_queue: false
-      single_instance: false
-    versions:
-      mac: ''
-      windows: iTunes.exe
-      linux:
+    - name: itunes
+      properties:
+        auto_queue: false
+        single_instance: false
+      versions:
+        mac: ''
+        windows: iTunes.exe
+        linux:
+    - name: iphoto
+      properties:
+        auto_queue: false
+        single_instance: false
+      versions:
+        mac: iPhoto
+        windows:
+        linux:
 status:
   applications:
-  - application: itunes
-    computers:
-    - computer: macbook-pro
-      timestamp:
-        started: 444
-        stopped: 0
-    next:
-  counter: 499
+    - application: itunes
+      computers:
+        - computer: macbook-pro
+          timestamp:
+            started: 444
+            stopped: 0
+      next:
+  counter: 501
